+++ website/src/routes/list-item/developer-experience.svelte
@@ -16,10 +16,11 @@
   } from '@uui';
   import type { ListItemCheckboxProps } from '@defense-unicorns/unicorn-ui/src/lib/List/ListItemCheckbox.types';
 
-  type ExtendedListItems1 = ListItemProps<EventTarget> & {
-    leadingAdornment?: boolean;
-    trailingAdornment?: boolean;
-  };
+  type ExtendedListItems1 = svelte.JSX.IntrinsicAttributes &
+    ListItemProps<EventTarget> & {
+      leadingAdornment?: boolean;
+      trailingAdornment?: boolean;
+    };
 
   const listItemsIteration1: ExtendedListItems1[] = [
     { text: 'List Item', selected: true },
@@ -48,10 +49,11 @@
     }
   ];
 
-  type ExtendedListItems2 = ListItemProps<EventTarget> & {
-    leadingAdornment?: Record<string, string>;
-    trailingAdornment?: ComponentAsProp<IconButtonProps | ListItemCheckboxProps>;
-  };
+  type ExtendedListItems2 = svelte.JSX.IntrinsicAttributes &
+    ListItemProps<EventTarget> & {
+      leadingAdornment?: Record<string, any>;
+      trailingAdornment?: Record<string, any>;
+    };
 
   const listItemsIteration2: ExtendedListItems2[] = [
     { text: 'List Item', secondaryText: 'Secondary Text' },